Querying Particle Data from Database
The MySQL database has been attached to the particle site, and particle data is pulled from each unique user to access the site and stored in the database.

On the python client side, you can now access the database and pull the coord data.

***NOTE***
To access the database, I need to add your home ip to the remote connection authorization.
Until I do this, you will get an access denied error when you try to run the code.